+++ packages/frontend/src/components/TableConfigPanel/ColumnConfiguration.tsx
@@ -1,9 +1,11 @@
-import { Button, InputGroup } from '@blueprintjs/core';
+import { Button, Colors, Icon, InputGroup } from '@blueprintjs/core';
 import React from 'react';
 import { useVisualizationContext } from '../LightdashVisualization/VisualizationProvider';
 import {
     ColumnConfigurationWrapper,
     ColumnWrapper,
+    EyeButton,
+    FrozenIcon,
 } from './ColumnConfiguration.styles';
 
 export const ColumnConfiguration: React.FC = () => {
@@ -15,6 +17,7 @@ export const ColumnConfiguration: React.FC = () => {
             getHeader,
             getDefaultColumnLabel,
             isColumnVisible,
+            isColumnFrozen,
         },
     } = useVisualizationContext();
     return (
@@ -36,7 +39,7 @@ export const ColumnConfiguration: React.FC = () => {
 
                         {(!pivotDimensions ||
                             !pivotDimensions.includes(fieldId)) && (
-                            <Button
+                            <EyeButton
                                 icon={
                                     isColumnVisible(fieldId)
                                         ? 'eye-off'
@@ -45,6 +48,23 @@ export const ColumnConfiguration: React.FC = () => {
                                 onClick={() => {
                                     updateColumnProperty(fieldId, {
                                         visible: !isColumnVisible(fieldId),
+                                    });
+                                }}
+                            />
+                        )}
+                        {pivotDimensions === undefined && (
+                            <FrozenIcon
+                                icon={
+                                    isColumnFrozen(fieldId) ? 'lock' : 'unlock'
+                                }
+                                color={
+                                    isColumnFrozen(fieldId)
+                                        ? 'black'
+                                        : Colors.GRAY5
+                                }
+                                onClick={() => {
+                                    updateColumnProperty(fieldId, {
+                                        frozen: !isColumnFrozen(fieldId),
                                     });
                                 }}
                             />