+++ packages/frontend/src/components/common/Table/index.tsx
@@ -23,6 +23,7 @@ const ResultsTable: FC<Props> = ({
     emptyState,
     hideRowNumbers,
     className,
+    showColumnCalculation,
     ...rest
 }) => {
     const LoadingState = loadingState || States.LoadingState;
@@ -30,7 +31,11 @@ const ResultsTable: FC<Props> = ({
     const EmptyState = emptyState || States.EmptyState;
 
     return (
-        <TableProvider hideRowNumbers={hideRowNumbers} {...rest}>
+        <TableProvider
+            hideRowNumbers={hideRowNumbers}
+            showColumnCalculation={showColumnCalculation}
+            {...rest}
+        >
             <TableContainer
                 className={`cohere-block${className ? ` ${className}` : ''}`}
                 $shouldExpand={$shouldExpand}