+++ packages/frontend/src/hooks/tableVisualization/useTableConfig.ts
@@ -91,7 +91,10 @@ const useTableConfig = (
         (fieldId: string) => columnProperties[fieldId]?.visible ?? true,
         [columnProperties],
     );
-
+    const isColumnFrozen = useCallback(
+        (fieldId: string) => columnProperties[fieldId]?.frozen === true,
+        [columnProperties],
+    );
     const getHeader = useCallback(
         (fieldId: string) => {
             return columnProperties[fieldId]?.name;
@@ -128,6 +131,7 @@ const useTableConfig = (
                 isColumnVisible,
                 showTableNames,
                 getHeader,
+                isColumnFrozen,
             });
         }
     }, [
@@ -204,6 +208,7 @@ const useTableConfig = (
         getHeader,
         getDefaultColumnLabel,
         isColumnVisible,
+        isColumnFrozen,
     };
 };