+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [0.315.0](https://github.com/lightdash/lightdash/compare/0.314.0...0.315.0) (2022-11-07)
+
+
+### Features
+
+* freeze table columns ([#3645](https://github.com/lightdash/lightdash/issues/3645)) ([2217c32](https://github.com/lightdash/lightdash/commit/2217c32aa3c2cedc93953a96ff0faecbc7df8bec))
+
+# [0.314.0](https://github.com/lightdash/lightdash/compare/0.313.4...0.314.0) (2022-11-07)
+
+
+### Features
+
+* remove magic metrics ([#3708](https://github.com/lightdash/lightdash/issues/3708)) ([95e9a4d](https://github.com/lightdash/lightdash/commit/95e9a4d0f24c18f017d6332c9229b08fe82df174))
+
 ## [0.313.4](https://github.com/lightdash/lightdash/compare/0.313.3...0.313.4) (2022-11-07)